Add poptorch cond check when use_encoder_true is enabled
